requirements.txt
pytest~=7.4.0
selene == 2.0.0rc4
allure-pytest>=2.13.2
python-dotenv==1.0.0